[BACKPORT 2024.1][PLAT-14624] skip deleting key pair from cloud in case skipKeyPairValidate is enabled
Summary:
skip deleting key pair from cloud in case skipKeyPairValidate is enabled
Depends on - D38818
Original Commit - https://github.com/yugabyte/yugabyte-db/commit/a2e9247ce177cd3c24a27c0929e4843994191c82

Test Plan:
Create AWS Provider with one region say Oregon with custom key and skipKeyPairValidate
Create a 3 node 3 rf universe in it
Edit provider to add a new region say Frankfurt
Perform full move to shift all the nodes to the newer region
Remove the older region (Oregon)

Reviewers: #yba-api-review!, amalyshev, yshchetinin, asharma

Reviewed By: asharma

Subscribers: yugaware, asharma

Differential Revision: https://phorge.dev.yugabyte.com/D38872

+++ managed/src/main/java/com/yugabyte/yw/controllers/handlers/RegionHandler.java
@@ -140,10 +140,14 @@ public class RegionHandler {
     if (provider.getCloudCode().equals(CloudType.aws)
         && provider.getAllAccessKeys() != null
         && provider.getAllAccessKeys().size() > 0) {
-      String keyPairName = AccessKey.getLatestKey(provider.getUuid()).getKeyCode();
-      log.info(
-          String.format("Deleting keyPair %s from region %s", keyPairName, regionUUID.toString()));
-      awsCloudImpl.deleteKeyPair(provider, region, keyPairName);
+      AccessKey key = AccessKey.getLatestKey(provider.getUuid());
+      if (!key.getKeyInfo().skipKeyValidateAndUpload) {
+        String keyPairName = key.getKeyCode();
+        log.info(
+            String.format(
+                "Deleting keyPair %s from region %s", keyPairName, regionUUID.toString()));
+        awsCloudImpl.deleteKeyPair(provider, region, keyPairName);
+      }
     }
     return region;
   }
